requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-genai
groq
requests
gunicorn
python-dotenv
ddgs
redis
easyocr
Pillow
numpy
